requirements.txt
asgiref==3.5.2
Django==4.0.7
psycopg2-binary==2.9.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
environs[django]==9.5.0
pillow==9.0.1
django-debug-toolbar==3.4.0
whitenoise==6.1.0
gunicorn==20.1.0
djangorestframework==3.13.0
django-storages[google]
django-tinymce==3.5.0
easy-thumbnails==2.8.3
django-cleanup==6.0.0
django-cors-headers~=3.10.0